+++ doc/source/changelog.rst
@@ -23,6 +23,7 @@ Master: 1.6.2
 * Lower priority of implicit conversion to temporary with initializer_list ctor
 * Add type reduction pythonization for trimming expression template type trees
 * Allow mixing std::string and str as dictionary keys
+* Support C-style pointer-to-struct as array
 * Remove '-isystem' from makepch flags
 * Extended documentation